+++ server/src/main/java/org/opensearch/search/aggregations/bucket/histogram/AutoDateHistogramAggregator.java
@@ -58,8 +58,8 @@ import org.opensearch.search.aggregations.bucket.terms.LongKeyedBucketOrds;
 import org.opensearch.search.aggregations.support.ValuesSource;
 import org.opensearch.search.aggregations.support.ValuesSourceConfig;
 import org.opensearch.search.internal.SearchContext;
-import org.opensearch.search.optimization.ranges.DateHistogramAggregatorBridge;
-import org.opensearch.search.optimization.ranges.OptimizationContext;
+import org.opensearch.search.optimization.filterrewrite.DateHistogramAggregatorBridge;
+import org.opensearch.search.optimization.filterrewrite.OptimizationContext;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -67,6 +67,8 @@ import java.util.Map;
 import java.util.function.BiConsumer;
 import java.util.function.Function;
 import java.util.function.LongToIntFunction;
+
+import static org.opensearch.search.optimization.filterrewrite.DateHistogramAggregatorBridge.segmentMatchAll;
 
 /**
  * An aggregator for date values that attempts to return a specific number of
@@ -198,14 +200,10 @@ abstract class AutoDateHistogramAggregator extends DeferableBucketAggregator {
             }
 
             @Override
-            protected Function<Object, Long> bucketOrdProducer() {
+            protected Function<Long, Long> bucketOrdProducer() {
                 return (key) -> getBucketOrds().add(0, preparedRounding.round((long) key));
             }
 
-            @Override
-            protected boolean segmentMatchAll(LeafReaderContext leaf) throws IOException {
-                return segmentMatchAll(context, leaf);
-            }
         });
         if (optimizationContext.canOptimize(parent, subAggregators.length, context)) {
             optimizationContext.prepare();
@@ -241,7 +239,7 @@ abstract class AutoDateHistogramAggregator extends DeferableBucketAggregator {
             return LeafBucketCollector.NO_OP_COLLECTOR;
         }
 
-        boolean optimized = optimizationContext.tryOptimize(ctx, this::incrementBucketDocCount);
+        boolean optimized = optimizationContext.tryOptimize(ctx, this::incrementBucketDocCount, segmentMatchAll(context, ctx));
         if (optimized) throw new CollectionTerminatedException();
 
         final SortedNumericDocValues values = valuesSource.longValues(ctx);
